Changing index creation workflow from "post-snapshot-import" phase of import-schema to default phase. (#1537)
Changing index creation workflow from "post-snapshot-import" phase of import-schema to default phase. 

**Reason:** 
Based on perf experiments, it was shown that pre-creating indexes and then loading data was faster as compared to loading data and backfilling indexes. 
The theory here is:
- Index creation is serial(parallel indexes cannot be created on YB as of now). Furthermore, as a result, there will be as many full table scans as there are no. of indexes 
- On the other hand, writing to indexes at the time of INSERT/UPDATE is done in parallel.
