+++ migtests/tests/mysql/indexes/validate
@@ -17,29 +17,12 @@ EXPECTED_ROW_COUNT = {
 	'single_index_test': 4
 }
 
-EXPECTED_INDEX_COUNT = {
-	'desc_index_test': 2,
-	'inunique_index_test': 3,
-	'mult_index_test': 2,
-	'outunique_index_test': 2,
-	'primary_index_test': 1,
-	'single_index_test': 2
-}
 
 def migration_completed_checks(tgt):
-	table_list = tgt.get_table_names("test_mysql_indexes")
-	print("table_list:", table_list)
-	assert len(table_list) == 6
-
 	got_row_count = tgt.row_count_of_all_tables("test_mysql_indexes")
 	for table_name, row_count in EXPECTED_ROW_COUNT.items():
 		print(f"table_name: {table_name}, row_count: {got_row_count[table_name]}")
 		assert row_count == got_row_count[table_name]
-
-	get_index_cnt = tgt.get_count_index_on_table("test_mysql_indexes")
-	for table_name, index_count in EXPECTED_INDEX_COUNT.items():
-		print(f"table_name: {table_name}, index_count: {get_index_cnt[table_name]}")
-		assert index_count == get_index_cnt[table_name]
 
 
 if __name__ == "__main__":
