+++ migtests/tests/pg/misc-objects-2/validate-schema
@@ -1,0 +1,27 @@
+#!/usr/bin/env python3
+
+import yb
+
+def main():
+	yb.run_checks(migration_completed_checks)
+
+
+#=============================================================================
+
+EXPECTED_TRIGGER_LIST = ["audit_trigger",]
+
+def migration_completed_checks(tgt):
+	table_list = tgt.get_table_names("public")
+	print("table_list:", table_list)
+	assert len(table_list) == 5 
+
+	fetched_triggers = tgt.fetch_all_triggers("public")
+	print(f"fetched triggers list: {fetched_triggers}\n Expected triggers: {EXPECTED_TRIGGER_LIST}")
+	assert fetched_triggers == EXPECTED_TRIGGER_LIST
+
+	fetched_procedures_functions = tgt.fetch_all_procedures("public")
+	print(f"count of fecthed procedures/functions - {len(fetched_procedures_functions)}")
+	assert len(fetched_procedures_functions) == 40
+
+if __name__ == "__main__":
+	main()
