+++ yb-voyager/cmd/importData.go
@@ -1080,7 +1080,7 @@ func beforeIndexCreation(sqlInfo sqlInfo, conn **pgx.Conn, objType string) error
 		return fmt.Errorf("extract qualified index name from DDL [%v]: %w", sqlInfo.stmt, err)
 	}
 	if invalidTargetIndexesCache == nil {
-		invalidTargetIndexesCache, err = tdb.InvalidIndexes()
+		invalidTargetIndexesCache, err = getInvalidIndexes(conn)
 		if err != nil {
 			return fmt.Errorf("failed to fetch invalid indexes: %w", err)
 		}
@@ -1098,6 +1098,32 @@ func beforeIndexCreation(sqlInfo sqlInfo, conn **pgx.Conn, objType string) error
 	// print the index name as index creation takes time and user can see the progress
 	color.Yellow("creating index %s ...", fullyQualifiedObjName)
 	return nil
+}
+
+func getInvalidIndexes(conn **pgx.Conn) (map[string]bool, error) {
+	var result = make(map[string]bool)
+	// NOTE: this shouldn't fetch any predefined indexes of pg_catalog schema (assuming they can't be invalid) or indexes of other successful migrations
+	query := "SELECT indexrelid::regclass FROM pg_index WHERE indisvalid = false"
+
+	rows, err := (*conn).Query(context.Background(), query)
+	if err != nil {
+		return nil, fmt.Errorf("querying invalid indexes: %w", err)
+	}
+	defer rows.Close()
+
+	for rows.Next() {
+		var fullyQualifiedIndexName string
+		err := rows.Scan(&fullyQualifiedIndexName)
+		if err != nil {
+			return nil, fmt.Errorf("scanning row for invalid index name: %w", err)
+		}
+		// if schema is not provided by catalog table, then it is public schema
+		if !strings.Contains(fullyQualifiedIndexName, ".") {
+			fullyQualifiedIndexName = fmt.Sprintf("public.%s", fullyQualifiedIndexName)
+		}
+		result[fullyQualifiedIndexName] = true
+	}
+	return result, nil
 }
 
 // TODO: This function is a duplicate of the one in tgtdb/yb.go. Consolidate the two.
