+++ yb-voyager/src/tgtdb/oracle.go
@@ -672,10 +672,6 @@ func (tdb *TargetOracleDB) isQueryResultNonEmpty(query string) bool {
 	return rows.Next()
 }
 
-func (tdb *TargetOracleDB) InvalidIndexes() (map[string]bool, error) {
-	return nil, nil
-}
-
 // this will be only called by FallForward or FallBack DBs
 func (tdb *TargetOracleDB) ClearMigrationState(migrationUUID uuid.UUID, exportDir string) error {
 	log.Infof("clearing migration state for migrationUUID: %s", migrationUUID)
