+++ yb-voyager/src/tgtdb/target_db_interface.go
@@ -49,7 +49,6 @@ type TargetDB interface {
 	EnableGeneratedAlwaysAsIdentityColumns(tableColumnsMap *utils.StructMap[sqlname.NameTuple, []string]) error
 	EnableGeneratedByDefaultAsIdentityColumns(tableColumnsMap *utils.StructMap[sqlname.NameTuple, []string]) error
 	ClearMigrationState(migrationUUID uuid.UUID, exportDir string) error
-	InvalidIndexes() (map[string]bool, error)
 	GetCallhomeTargetDBInfo() *callhome.TargetDBDetails
 	// NOTE: The following four methods should not be used for arbitrary query
 	// execution on TargetDB. The should be only used from higher level
